requirements.txt
apscheduler==3.10.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
datacollectors==0.7.7
entsoe-py==0.6.8
et-xmlfile==1.1.0
idna==3.7
incentivedkutils==0.1.3
lxml==5.2.2
numpy==1.26.4
openpyxl==3.1.3
pandas==2.2.2
pathlib==1.0.1
pycountry==24.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
requests==2.32.3
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
xmltodict==0.13.0
